restructuring for gradio interface; logging per tool
