--- Toolbox/Pages/__init__.py
@@ -1,5 +1,0 @@
-import os
-import sys
-
-# Add init real path to sys
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
