+++ Toolbox/api.py
@@ -1,34 +1,38 @@
 import gradio as gr
 
-from Toolbox.Pages.common import *
+from common import *
 
-from Toolbox.Tools.Points import points
+from Tools.API import api
 
 EXIT_APP = False
+log_file = "annotate.log"
 
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Module
 # ----------------------------------------------------------------------------------------------------------------------
 
-def module_callback(images, sample_method, num_points, output_dir):
+def module_callback(username, password, source_id_1, source_id_2, points, prefix, output_dir):
     """
 
     """
     console = sys.stdout
-    sys.stdout = Logger(LOG_PATH)
+    sys.stdout = Logger(log_file)
 
     args = argparse.Namespace(
-        images=images,
-        sample_method=sample_method,
-        num_points=num_points,
+        username=username,
+        password=password,
+        source_id_1=source_id_1,
+        source_id_2=source_id_2,
+        points=points,
+        prefix=prefix,
         output_dir=output_dir,
     )
 
     try:
         # Call the function
         gr.Info("Starting process...")
-        points(args)
+        api(args)
         print("\nDone.")
         gr.Info("Completed process!")
     except Exception as e:
@@ -57,23 +61,27 @@ def create_interface():
     """
 
     """
-    Logger(LOG_PATH).reset_logs()
+    logger = Logger(log_file)
+    logger.reset_logs()
 
-    with gr.Blocks(title="Points 🏓", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
+    with gr.Blocks(title="CoralNet API 🕹️", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
         # Title
-        gr.Markdown("# Points 🏓")
+        gr.Markdown("# CoralNet API 🕹️")
 
-        # Browse button
-        images = gr.Textbox(f"{DATA_DIR}", label="Selected Image Directory")
-        dir_button = gr.Button("Browse Directory")
-        dir_button.click(choose_directory, outputs=images, show_progress="hidden")
+        # Input Parameters
+        with gr.Row():
+            username = gr.Textbox(os.getenv('CORALNET_USERNAME'), label="Username", type='email')
+            password = gr.Textbox(os.getenv('CORALNET_PASSWORD'), label="Password", type='password')
 
         with gr.Row():
-            sample_method = gr.Dropdown(label="Sample Method",
-                                        choices=['Uniform', 'Random', 'Stratified'],
-                                        multiselect=False)
+            source_id_1 = gr.Textbox("", label="Source ID (for images)")
+            source_id_2 = gr.Textbox("", label="Source ID (for model)")
+            prefix = gr.Textbox("", label="Image Name Prefix")
 
-            num_points = gr.Number(label="Number of Points", precision=0)
+        # Files button
+        points = gr.Textbox("", label="Selected Points File")
+        files_button = gr.Button("Browse Files")
+        files_button.click(choose_files, outputs=points, show_progress="hidden")
 
         # Browse button
         output_dir = gr.Textbox(f"{DATA_DIR}", label="Selected Output Directory")
@@ -84,9 +92,12 @@ def create_interface():
             # Run button (callback)
             run_button = gr.Button("Run")
             run = run_button.click(module_callback,
-                                   [images,
-                                    sample_method,
-                                    num_points,
+                                   [username,
+                                    password,
+                                    source_id_1,
+                                    source_id_2,
+                                    points,
+                                    prefix,
                                     output_dir])
 
             stop_button = gr.Button(value="Stop")
@@ -95,7 +106,7 @@ def create_interface():
         with gr.Accordion("Console Logs"):
             # Add logs
             logs = gr.Code(label="", language="shell", interactive=False, container=True, lines=30)
-            interface.load(read_logs, None, logs, every=1)
+            interface.load(logger.read_logs, None, logs, every=1)
 
     interface.launch(prevent_thread_lock=True, server_port=get_port(), inbrowser=True, show_error=True)
 
@@ -116,5 +127,4 @@ except:
     pass
 
 finally:
-    Logger(LOG_PATH).reset_logs()
-
+    Logger(log_file).reset_logs()
