+++ Toolbox/labelset.py
@@ -1,10 +1,11 @@
 import gradio as gr
 
-from Toolbox.Pages.common import *
+from common import *
 
-from Toolbox.Tools.Labelset import labelset
+from Tools.Labelset import labelset
 
 EXIT_APP = False
+log_file = "labelset.log"
 
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -16,7 +17,7 @@ def module_callback(username, password, source_id, labelset_name, short_code, fu
 
     """
     console = sys.stdout
-    sys.stdout = Logger(LOG_PATH)
+    sys.stdout = Logger(log_file)
 
     args = argparse.Namespace(
         username=username,
@@ -62,7 +63,8 @@ def create_interface():
     """
 
     """
-    Logger(LOG_PATH).reset_logs()
+    logger = Logger(log_file)
+    logger.reset_logs()
 
     with gr.Blocks(title="Labelset 📝", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
         # Title
@@ -108,7 +110,7 @@ def create_interface():
         with gr.Accordion("Console Logs"):
             # Add logs
             logs = gr.Code(label="", language="shell", interactive=False, container=True, lines=30)
-            interface.load(read_logs, None, logs, every=1)
+            interface.load(logger.read_logs, None, logs, every=1)
 
     interface.launch(prevent_thread_lock=True, server_port=get_port(), inbrowser=True, show_error=True)
 
@@ -129,5 +131,4 @@ except:
     pass
 
 finally:
-    Logger(LOG_PATH).reset_logs()
-
+    Logger(log_file).reset_logs()
