+++ Toolbox/segmentation.py
@@ -1,44 +1,43 @@
 import gradio as gr
 
-from Toolbox.Pages.common import *
+from common import *
 
-from Toolbox.Tools.Classification import classification
-from Toolbox.Tools.Classification import get_classifier_losses
-from Toolbox.Tools.Classification import get_classifier_metrics
-from Toolbox.Tools.Classification import get_classifier_encoders
-from Toolbox.Tools.Classification import get_classifier_optimizers
+from Tools.Segmentation import segmentation
+from Tools.Segmentation import get_segmentation_encoders
+from Tools.Segmentation import get_segmentation_decoders
+from Tools.Segmentation import get_segmentation_metrics
+from Tools.Segmentation import get_segmentation_losses
+from Tools.Segmentation import get_segmentation_optimizers
 
 EXIT_APP = False
+log_file = "segmentation.log"
 
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Module
 # ----------------------------------------------------------------------------------------------------------------------
 
-def module_callback(patches, pre_trained_path, encoder_name, freeze_encoder, loss_function, weighted_loss, metrics,
-                    optimizer, learning_rate, augment_data, dropout_rate, num_epochs, batch_size, tensorboard,
+def module_callback(masks, color_map, pre_trained_path, encoder_name, decoder_name, metrics, loss_function,
+                    freeze_encoder, optimizer, learning_rate, augment_data, num_epochs, batch_size, tensorboard,
                     output_dir):
     """
 
     """
     console = sys.stdout
-    sys.stdout = Logger(LOG_PATH)
-
-    # Custom pre-processing
-    patches = patches.split(" ")
+    sys.stdout = Logger(log_file)
 
     args = argparse.Namespace(
-        patches=patches,
+        masks=masks,
+        color_map=color_map,
         pre_trained_path=pre_trained_path,
         encoder_name=encoder_name,
-        freeze_encoder=freeze_encoder,
-        loss_function=loss_function,
-        weighted_loss=weighted_loss,
+        decoder_name=decoder_name,
         metrics=metrics,
+        loss_function=loss_function,
+        freeze_encoder=freeze_encoder,
         optimizer=optimizer,
         learning_rate=learning_rate,
         augment_data=augment_data,
-        dropout_rate=dropout_rate,
         num_epochs=num_epochs,
         batch_size=batch_size,
         tensorboard=tensorboard,
@@ -48,7 +47,7 @@ def module_callback(patches, pre_trained_path, encoder_name, freeze_encoder, los
     try:
         # Call the function
         gr.Info("Starting process...")
-        classification(args)
+        segmentation(args)
         print("\nDone.")
         gr.Info("Completed process!")
     except Exception as e:
@@ -86,60 +85,61 @@ def create_interface():
     """
 
     """
-    Logger(LOG_PATH).reset_logs()
+    logger = Logger(log_file)
+    logger.reset_logs()
 
     with gr.Blocks(title="Train 👨‍💻", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
         # Title
-        gr.Markdown("Train 👨‍💻")
+        gr.Markdown("# Train 👨‍💻")
 
         # Browse button
-        patches = gr.Textbox(label="Selected Patches File")
-        file_button = gr.Button("Browse Files")
-        file_button.click(choose_files, outputs=patches, show_progress="hidden")
+        masks = gr.Textbox(label="Selected Masks File")
+        files_button = gr.Button("Browse Files")
+        files_button.click(choose_file, outputs=masks, show_progress="hidden")
+
+        color_map = gr.Textbox(label="Selected Color Map File")
+        files_button = gr.Button("Browse Files")
+        files_button.click(choose_file, outputs=color_map, show_progress="hidden")
 
         pre_trained_path = gr.Textbox(label="Selected Pre-trained Encoder File")
         file_button = gr.Button("Browse Files")
         file_button.click(choose_files, outputs=pre_trained_path, show_progress="hidden")
 
-        with gr.Group("Training Parameters"):
-            #
-            with gr.Row():
-                encoder_name = gr.Dropdown(label="Encoder", multiselect=False, allow_custom_value=False,
-                                           choices=get_classifier_encoders())
+        with gr.Row():
+            encoder_name = gr.Dropdown(label="Encoder", multiselect=False, allow_custom_value=False,
+                                       choices=get_segmentation_encoders())
 
-                freeze_encoder = gr.Slider(0.0, label="Freeze Encoder", minimum=0.0, maximum=1.0, step=0.01)
+            freeze_encoder = gr.Slider(0.0, label="Freeze Encoder", minimum=0.0, maximum=1.0, step=0.01)
 
-            with gr.Row():
-                optimizer = gr.Dropdown(label="Optimizer", multiselect=False, allow_custom_value=False,
-                                        choices=get_classifier_optimizers())
+            decoder_name = gr.Dropdown(label="Decoder", multiselect=False, allow_custom_value=False,
+                                       choices=get_segmentation_decoders())
 
-                learning_rate = gr.Slider(0.0001, label="Initial Learning Rate",
-                                          minimum=0.00001, maximum=1, step=0.0001)
+        with gr.Row():
+            optimizer = gr.Dropdown(label="Optimizer", multiselect=False, allow_custom_value=False,
+                                    choices=get_segmentation_optimizers())
 
-            with gr.Row():
-                metrics = gr.Dropdown(label="Metrics", multiselect=True, allow_custom_value=False,
-                                      choices=get_classifier_metrics())
+            learning_rate = gr.Slider(0.0001, label="Initial Learning Rate",
+                                      minimum=0.00001, maximum=1, step=0.0001)
 
-                loss_function = gr.Dropdown(label="Loss Function", multiselect=False, allow_custom_value=False,
-                                            choices=get_classifier_losses())
+        with gr.Row():
+            metrics = gr.Dropdown(label="Metrics", multiselect=True, allow_custom_value=False,
+                                  choices=get_segmentation_metrics())
 
-                weighted_loss = gr.Dropdown(label="Weighted Loss", multiselect=False, allow_custom_value=False,
-                                            choices=[True, False])
+            loss_function = gr.Dropdown(label="Loss Function", multiselect=False, allow_custom_value=False,
+                                        choices=get_segmentation_losses())
 
-            with gr.Row():
-                augment_data = gr.Dropdown(label="Augment Data", multiselect=False, allow_custom_value=False,
-                                           choices=[True, False])
+            augment_data = gr.Dropdown(label="Augment Data", multiselect=False, allow_custom_value=False,
+                                       choices=[True, False])
 
-                dropout_rate = gr.Slider(label="Dropout Rate", minimum=0, maximum=1, step=0.1)
+        with gr.Row():
+            num_epochs = gr.Number(25, label="Number of Epochs", precision=0)
 
-            with gr.Row():
-                num_epochs = gr.Number(25, label="Number of Epochs", precision=0)
+            batch_size = gr.Number(8, label="Batch Size (Power of 2 Recommended)", precision=0)
 
-                batch_size = gr.Number(128, label="Batch Size (Power of 2 Recommended)", precision=0)
+            tensorboard = gr.Dropdown(label="Tensorboard", multiselect=False, allow_custom_value=False,
+                                      choices=[True, False])
 
-                tensorboard = gr.Dropdown(label="Tensorboard", multiselect=False, allow_custom_value=False,
-                                          choices=[True, False])
-
+        # Browse button
         output_dir = gr.Textbox(f"{DATA_DIR}", label="Selected Output Directory")
         dir_button = gr.Button("Browse Directory")
         dir_button.click(choose_directory, outputs=output_dir, show_progress="hidden")
@@ -148,17 +148,17 @@ def create_interface():
             # Run button (callback)
             run_button = gr.Button("Run")
             run = run_button.click(module_callback,
-                                   [patches,
+                                   [masks,
+                                    color_map,
                                     pre_trained_path,
                                     encoder_name,
-                                    freeze_encoder,
-                                    loss_function,
-                                    weighted_loss,
+                                    decoder_name,
                                     metrics,
+                                    loss_function,
+                                    freeze_encoder,
                                     optimizer,
                                     learning_rate,
                                     augment_data,
-                                    dropout_rate,
                                     num_epochs,
                                     batch_size,
                                     tensorboard,
@@ -168,9 +168,9 @@ def create_interface():
             stop = stop_button.click(exit_interface)
 
         with gr.Accordion("Console Logs"):
-            # Add to console in page
+            # Add logs
             logs = gr.Code(label="", language="shell", interactive=False, container=True, lines=30)
-            interface.load(read_logs, None, logs, every=1)
+            interface.load(logger.read_logs, None, logs, every=1)
 
         with gr.Accordion("TensorBoard"):
             # Display Tensorboard in page
@@ -197,4 +197,4 @@ except:
     pass
 
 finally:
-    Logger(LOG_PATH).reset_logs()
+    Logger(log_file).reset_logs()
