+++ Toolbox/segmentation_3d.py
@@ -1,10 +1,11 @@
 import gradio as gr
 
-from Toolbox.Pages.common import *
+from common import *
 
-from Toolbox.Tools.Segmentation3D import segmentation3d
+from Tools.Segmentation3D import segmentation3d
 
 EXIT_APP = False
+log_file = "segmentation_3d.log"
 
 
 # ----------------------------------------------------------------------------------------------------------------------
@@ -16,7 +17,7 @@ def module_callback(metashape_license, project_file, masks_file, color_map, mask
 
     """
     console = sys.stdout
-    sys.stdout = Logger(LOG_PATH)
+    sys.stdout = Logger(log_file)
 
     args = argparse.Namespace(
         metashape_license=metashape_license,
@@ -60,7 +61,8 @@ def create_interface():
     """
 
     """
-    Logger(LOG_PATH).reset_logs()
+    logger = Logger(log_file)
+    logger.reset_logs()
 
     with gr.Blocks(title="3D Semantic Segmentation 🤖️", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
         # Title
@@ -108,7 +110,7 @@ def create_interface():
         with gr.Accordion("Console Logs"):
             # Add logs
             logs = gr.Code(label="", language="shell", interactive=False, container=True, lines=30)
-            interface.load(read_logs, None, logs, every=1)
+            interface.load(logger.read_logs, None, logs, every=1)
 
     interface.launch(prevent_thread_lock=True, server_port=get_port(), inbrowser=True, show_error=True)
 
@@ -129,6 +131,4 @@ except:
     pass
 
 finally:
-    Logger(LOG_PATH).reset_logs()
-
-
+    Logger(log_file).reset_logs()
