+++ Toolbox/upload.py
@@ -1,42 +1,40 @@
 import gradio as gr
 
-from Toolbox.Pages.common import *
+from common import *
 
-from Toolbox.Tools.Download import download
-from Toolbox.Tools.Download import get_updated_labelset_list
+from Tools.Upload import upload
 
 EXIT_APP = False
+log_file = "upload.log"
 
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Module
 # ----------------------------------------------------------------------------------------------------------------------
 
-def module_callback(username, password, source_ids, source_df, labelset_df, sources_with, output_dir, headless):
+def module_callback(username, password, source_id, images, prefix, annotations, labelset, headless, output_dir):
     """
 
     """
     console = sys.stdout
-    sys.stdout = Logger(LOG_PATH)
-
-    # Custom pre-processing
-    source_ids = [str(id.strip()) for id in source_ids.split(" ") if id.strip()]
+    sys.stdout = Logger(log_file)
 
     args = argparse.Namespace(
         username=username,
         password=password,
-        source_ids=source_ids,
-        source_df=source_df,
-        labelset_df=labelset_df,
-        sources_with=sources_with,
-        output_dir=output_dir,
-        headless=headless
+        source_id=source_id,
+        images=images,
+        prefix=prefix,
+        annotations=annotations,
+        labelset=labelset,
+        headless=headless,
+        output_dir=output_dir
     )
 
     try:
         # Call the function
         gr.Info("Starting process...")
-        download(args)
+        upload(args)
         print("\nDone.")
         gr.Info("Completed process!")
     except Exception as e:
@@ -65,30 +63,34 @@ def create_interface():
     """
 
     """
-    Logger(LOG_PATH).reset_logs()
+    logger = Logger(log_file)
+    logger.reset_logs()
 
-    with gr.Blocks(title="CoralNet Download ⬇️", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
+    with gr.Blocks(title="Upload ⬆️", analytics_enabled=False, theme=gr.themes.Soft(), js=js) as interface:
         # Title
-        gr.Markdown("# CoralNet Downloader ⬇️")
+        gr.Markdown("# Upload ⬆️")
 
-        # Input Parameters
-        with gr.Tab("Download Source Data"):
-            with gr.Row():
-                username = gr.Textbox(os.getenv('CORALNET_USERNAME'), label="Username", type='email')
-                password = gr.Textbox(os.getenv('CORALNET_PASSWORD'), label="Password", type='password')
+        with gr.Row():
+            username = gr.Textbox(os.getenv('CORALNET_USERNAME'), label="Username", type='email')
+            password = gr.Textbox(os.getenv('CORALNET_PASSWORD'), label="Password", type='password')
 
-            with gr.Row():
-                source_ids = gr.Textbox("", label="Source IDs (space-separated)")
-                headless = gr.Checkbox(label="Run Browser in Headless Mode", value=True)
+        with gr.Row():
+            source_id = gr.Number(label="Source ID", precision=0)
+            prefix = gr.Textbox(label="Image Name Prefix")
+            headless = gr.Checkbox(label="Run Browser in Headless Mode", value=True)
 
-        with gr.Tab("Download CoralNet Dataframes"):
-            with gr.Row():
-                source_df = gr.Checkbox(label="Download Source DataFrame")
-                labelset_df = gr.Checkbox(label="Download Labelset DataFrame")
+        # Browse button
+        images = gr.Textbox(f"{DATA_DIR}", label="Selected Image Directory")
+        dir_button = gr.Button("Browse Directory")
+        dir_button.click(choose_directory, outputs=images, show_progress="hidden")
 
-            sources_with = gr.Dropdown(label="Sources with Labelsets",
-                                       choices=get_updated_labelset_list(),
-                                       multiselect=True)
+        annotations = gr.Textbox(label="Selected Annotation File")
+        file_button = gr.Button("Browse Files")
+        file_button.click(choose_file, outputs=annotations, show_progress="hidden")
+
+        labelset = gr.Textbox(label="Selected Labelset File")
+        file_button = gr.Button("Browse Files")
+        file_button.click(choose_file, outputs=labelset, show_progress="hidden")
 
         # Browse button
         output_dir = gr.Textbox(f"{DATA_DIR}", label="Selected Output Directory")
@@ -101,12 +103,13 @@ def create_interface():
             run = run_button.click(module_callback,
                                    [username,
                                     password,
-                                    source_ids,
-                                    source_df,
-                                    labelset_df,
-                                    sources_with,
-                                    output_dir,
-                                    headless])
+                                    source_id,
+                                    images,
+                                    prefix,
+                                    annotations,
+                                    labelset,
+                                    headless,
+                                    output_dir])
 
             stop_button = gr.Button(value="Stop")
             stop = stop_button.click(exit_interface)
@@ -114,7 +117,7 @@ def create_interface():
         with gr.Accordion("Console Logs"):
             # Add logs
             logs = gr.Code(label="", language="shell", interactive=False, container=True, lines=30)
-            interface.load(read_logs, None, logs, every=1)
+            interface.load(logger.read_logs, None, logs, every=1)
 
     interface.launch(prevent_thread_lock=True, server_port=get_port(), inbrowser=True, show_error=True)
 
@@ -135,6 +138,4 @@ except:
     pass
 
 finally:
-    Logger(LOG_PATH).reset_logs()
-
-
+    Logger(log_file).reset_logs()
